SubdomainHijacking/requirements.txt
requests
dnspython
stem
pysocks
